Made changes/fixes to history.
Made the length part of the history struct itself, so that if, for example, the history length could be set as an environment variable in the future, the history can keep track of it. Also fixed a bug with executing the last command in the history; previously, removing the command and adding it to the history/executing it was causing problems, so simply had to move around where the command is added to history and where it is freed.