Fixed problem with '.' being implicitly added to path.
My solution to name collisions was to simply always try the full path first when doing execv, but that had the side-effect of automatically executing anything in the directory with the same name automatically, so I simply changed the if statement to follow execvp rules; only execute the path by itself if it contains a slash. Also added some comments.